new: create ``mockify.core`` module
Currently, ``from mockify import ...`` is used to import core library
functionalities. This behaviour will be deprecated and removed in one of
upcoming releases in favour of dedicated ``core`` module. That will make
it a bit more efficient in terms of importing the library.

So instead of::

    from mockify import satisfied

You will have to do::

    from mockify.core import satisfied